fix: datatables loading issue
we're using bodytag to add the class mandant. but the extension
datatables_view also adds a class to it which is important for it to
work
our extension overwrites that class and therefore we need to move the
mandant class evwen higher.
we're now doing htmltag which is not the best way since we're
overwriting the whole htmltag not just adding to it but for now this
seems like the best options. need to be more careful though when CKAN
changes the htmltag implementation